requirements.txt
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
pyspeedtest==1.2.7
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
pyxdg==0.26
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
rope==0.17.0
SecretStorage==3.1.2
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
speedtest-cli==2.1.2
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
ujson==1.35
urllib3==1.25.9
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
wurlitzer==2.0.1
yapf==0.30.0
zipp==3.1.0
